+++ cg_lims/get/udfs.py
@@ -2,6 +2,10 @@ from genologics.lims import Lims
 from typing import Optional
 from datetime import date
 
+from cg_lims.EPPs.mongo.prep_microbial import LOG
+from cg_lims.exceptions import MissingArtifactError
+from cg_lims.get.artifacts import get_latest_artifact
+
 
 def get_udf_type(lims: Lims, udf_name: str, attach_to_name: str) -> Optional:
     """Get udf type.
@@ -19,3 +23,29 @@ def get_udf_type(lims: Lims, udf_name: str, attach_to_name: str) -> Optional:
     udf_type = udf_configs[0].root.attrib["type"]
 
     return udf_types[udf_type]
+
+
+def filter_process_udfs_by_model(lims: Lims, sample_id: str, process_type: str, model):
+    """"""
+    try:
+        artifact = get_latest_artifact(lims=lims, sample_id=sample_id, process_type=[process_type])
+    except MissingArtifactError as e:
+        LOG.info(e.message)
+        return dict()
+    process_udfs = dict(artifact.parent_process.udf.items())
+    udf_model = model(**process_udfs)
+    return udf_model.dict(exclude_none=True)
+
+
+def filter_process_artifact_udfs_by_model(
+    lims: Lims, sample_id: str, process_type: str, model
+) -> dict:
+    """"""
+    try:
+        artifact = get_latest_artifact(lims=lims, sample_id=sample_id, process_type=[process_type])
+    except MissingArtifactError as e:
+        LOG.info(e.message)
+        return dict()
+    artifact_udfs = dict(artifact.udf.items())
+    udf_model = model(**artifact_udfs)
+    return udf_model.dict(exclude_none=True)
